correzione bug

+++ DefaultSerial1RelayFraschetta.h
@@ -6,7 +6,7 @@
 class DefaultSerial1RelayF:DefaultSerial1F,public BasicSerialRelayF{
 protected: void Send(String s)override{DefaultSerial1F::SendLn(s);}
 public:
- DefaultSerial1RelayF(String Tag,uint8_t Pin,DigitalType Type=NormalLogic,MemorizeStatus Save=DoNotSaveStatus,uint16_t Baud=9600):BasicSerialRelayF(Tag,Pin,Type,Save),DefaultSerial1F(Baud){}
+ DefaultSerial1RelayF(String Tag,uint8_t Pin,DigitalType Type=NormalLogic,MemorizeStatus Save=DoNotSaveStatus,long Baud=9600):BasicSerialRelayF(Tag,Pin,Type,Save),DefaultSerial1F(Baud){}
 void Begin(){DefaultSerial1F::Begin(); BasicSerialRelayF::Begin();}
 };
 #endif
