+++ DefaultSerialRelayFraschetta.h
@@ -5,7 +5,7 @@
 class DefaultSerialRelayF:DefaultSerialF,public BasicSerialRelayF{
 protected: void Send(String s)override{DefaultSerialF::SendLn(s);}
 public:
- DefaultSerialRelayF(String Tag,uint8_t Pin,DigitalType Type=NormalLogic,MemorizeStatus Save=DoNotSaveStatus,uint16_t Baud=9600):BasicSerialRelayF(Tag,Pin,Type,Save),DefaultSerialF(Baud){}
+ DefaultSerialRelayF(String Tag,uint8_t Pin,DigitalType Type=NormalLogic,MemorizeStatus Save=DoNotSaveStatus,long Baud=9600):BasicSerialRelayF(Tag,Pin,Type,Save),DefaultSerialF(Baud){}
  void Begin(){DefaultSerialF::Begin(); BasicSerialRelayF::Begin();}
 };
 #endif
